add fluent api to regex assets
