+++ great_expectations/datasource/fluent/file_path_data_asset.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import copy
 import logging
-import re
 from pprint import pformat as pf
 from typing import (
     TYPE_CHECKING,
@@ -22,10 +21,6 @@ from great_expectations._docs_decorators import public_api
 from great_expectations.compatibility import pydantic
 from great_expectations.compatibility.typing_extensions import override
 from great_expectations.core.partitioners import (
-    PartitionerDaily,
-    PartitionerMonthly,
-    PartitionerPath,
-    PartitionerYearly,
     RegexPartitioner,
 )
 from great_expectations.datasource.fluent.batch_request import (
@@ -50,7 +45,6 @@ from great_expectations.datasource.fluent.interfaces import (
 if TYPE_CHECKING:
     from great_expectations.alias_types import PathStr
     from great_expectations.core.batch import BatchMarkers, LegacyBatchDefinition
-    from great_expectations.core.batch_definition import BatchDefinition
     from great_expectations.core.id_dict import BatchSpec
     from great_expectations.datasource.fluent.interfaces import (
         BatchMetadata,
@@ -62,40 +56,6 @@ if TYPE_CHECKING:
     )
 
 logger = logging.getLogger(__name__)
-
-
-class RegexMissingRequiredGroupsError(ValueError):
-    def __init__(self, missing_groups: set[str]):
-        message = (
-            "The following group(s) are required but are "
-            f"missing from the regex: {', '.join(missing_groups)}"
-        )
-        super().__init__(message)
-        self.missing_groups = missing_groups
-
-
-class RegexUnknownGroupsError(ValueError):
-    def __init__(self, unknown_groups: set[str]):
-        message = (
-            "Regex has the following group(s) which do not match "
-            f"batch parameters: {', '.join(unknown_groups)}"
-        )
-        super().__init__(message)
-        self.unknown_groups = unknown_groups
-
-
-class PathNotFoundError(ValueError):
-    def __init__(self, path: PathStr):
-        message = f"Provided path was not able to be resolved: {path} "
-        super().__init__(message)
-        self.path = path
-
-
-class AmbiguousPathError(ValueError):
-    def __init__(self, path: PathStr):
-        message = f"Provided path matched multiple targets, and must match exactly one: {path} "
-        super().__init__(message)
-        self.path = path
 
 
 class _FilePathDataAsset(DataAsset[DatasourceT, RegexPartitioner], Generic[DatasourceT]):
@@ -227,102 +187,6 @@ class _FilePathDataAsset(DataAsset[DatasourceT, RegexPartitioner], Generic[Datas
             batch_slice=batch_slice,
             partitioner=partitioner,
         )
-
-    def _add_batch_definition_path(self, name: str, path: PathStr) -> BatchDefinition:
-        """Adds a BatchDefinition which matches a single Path.
-
-        Parameters:
-            name: BatchDefinition name
-            path: File path relative to the Asset
-
-        Raises:
-             PathNotFoundError: path cannot be resolved
-             AmbiguousPathError: path matches more than one file
-        """
-        regex = re.compile(str(path))
-        matched_data_references = len(self._data_connector.get_matched_data_references(regex=regex))
-        # we require path to match exactly 1 file
-        if matched_data_references < 1:
-            raise PathNotFoundError(path=path)
-        elif matched_data_references > 1:
-            raise AmbiguousPathError(path=path)
-        return self.add_batch_definition(
-            name=name,
-            partitioner=PartitionerPath(regex=regex),
-        )
-
-    def _add_batch_definition_yearly(self, name: str, regex: re.Pattern) -> BatchDefinition:
-        """Adds a BatchDefinition which defines yearly batches by file name.
-
-        Parameters:
-            name: BatchDefinition name
-            regex: Regular Expression used to define batches by file name.
-                Must contain a single group `year`
-
-        Raises:
-            RegexMissingRequiredGroupsError: regex is missing the group `year`
-            RegexUnknownGroupsError: regex has groups other than `year`
-        """
-        REQUIRED_GROUP_NAME = {"year"}
-        self._assert_group_names_in_regex(regex=regex, required_group_names=REQUIRED_GROUP_NAME)
-        return self.add_batch_definition(
-            name=name,
-            partitioner=PartitionerYearly(regex=regex),
-        )
-
-    def _add_batch_definition_monthly(self, name: str, regex: re.Pattern) -> BatchDefinition:
-        """Adds a BatchDefinition which defines monthly batches by file name.
-
-        Parameters:
-            name: BatchDefinition name
-            regex: Regular Expression used to define batches by file name.
-                Must contain the groups `year` and `month`.
-
-        Raises:
-            RegexMissingRequiredGroupsError: regex is missing the groups `year` and/or `month`.
-            RegexUnknownGroupsError: regex has groups other than `year` and/or `month`.
-        """
-        REQUIRED_GROUP_NAMES = {"year", "month"}
-        self._assert_group_names_in_regex(regex=regex, required_group_names=REQUIRED_GROUP_NAMES)
-        return self.add_batch_definition(
-            name=name,
-            partitioner=PartitionerMonthly(regex=regex),
-        )
-
-    def _add_batch_definition_daily(self, name: str, regex: re.Pattern) -> BatchDefinition:
-        """Adds a BatchDefinition which defines daily batches by file name.
-
-        Parameters:
-            name: BatchDefinition name
-            regex: Regular Expression used to define batches by file name.
-                Must contain the groups `year`, `month`, and `day`.
-
-        Raises:
-            RegexMissingRequiredGroupsError: regex is missing the
-                groups `year`, `month`, and/or `day`.
-            RegexUnknownGroupsError: regex has groups other than `year`, `month`, and/or `day`.
-        """
-        REQUIRED_GROUP_NAMES = {"year", "month", "day"}
-        self._assert_group_names_in_regex(regex=regex, required_group_names=REQUIRED_GROUP_NAMES)
-        return self.add_batch_definition(
-            name=name,
-            partitioner=PartitionerDaily(regex=regex),
-        )
-
-    @classmethod
-    def _assert_group_names_in_regex(
-        cls, regex: re.Pattern, required_group_names: set[str]
-    ) -> None:
-        regex_parser = RegExParser(
-            regex_pattern=regex,
-        )
-        actual_group_names = set(regex_parser.group_names())
-        if not required_group_names.issubset(actual_group_names):
-            missing_groups = required_group_names - actual_group_names
-            raise RegexMissingRequiredGroupsError(missing_groups)
-        if not actual_group_names.issubset(required_group_names):
-            unknown_groups = actual_group_names - required_group_names
-            raise RegexUnknownGroupsError(unknown_groups)
 
     @override
     def _validate_batch_request(self, batch_request: BatchRequest) -> None:
